+++ ApiAdminShop.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.10.34916.146
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApiAdminShop", "ApiAdminShop\ApiAdminShop.csproj", "{5DC05211-37E3-4C7A-A608-E7D99D09A49B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data", "Data\Data.csproj", "{C788DB86-38B3-4D36-9E4A-9BF5DC7DE794}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{4215024B-3BDF-4EE9-86F5-32D76C0EC0DF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{5DC05211-37E3-4C7A-A608-E7D99D09A49B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5DC05211-37E3-4C7A-A608-E7D99D09A49B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5DC05211-37E3-4C7A-A608-E7D99D09A49B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C788DB86-38B3-4D36-9E4A-9BF5DC7DE794}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C788DB86-38B3-4D36-9E4A-9BF5DC7DE794}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C788DB86-38B3-4D36-9E4A-9BF5DC7DE794}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C788DB86-38B3-4D36-9E4A-9BF5DC7DE794}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4215024B-3BDF-4EE9-86F5-32D76C0EC0DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4215024B-3BDF-4EE9-86F5-32D76C0EC0DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4215024B-3BDF-4EE9-86F5-32D76C0EC0DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4215024B-3BDF-4EE9-86F5-32D76C0EC0DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
