--- Data/AppDbContext.cs
@@ -1,13 +1,0 @@
-﻿using Data.Entities;
-using Microsoft.EntityFrameworkCore;
-
-namespace Data;
-
-public class AppDbContext(DbContextOptions<AppDbContext> options):DbContext(options)
-{
-    public DbSet<Category> Categories { get; set; }
-    public DbSet<Product> Products { get; set; }
-    public DbSet<Order> Orders { get; set; }
-    public DbSet<OrderDetail> OrderDetails { get; set; }
-    public DbSet<User> Users { get; set; }
-}
