--- Data/Entities/Category.cs
@@ -1,10 +1,0 @@
-﻿namespace Data.Entities
-{
-    public  class Category : BaseEntity
-    {
-        public string Name {  get; set; } = string.Empty;
-        public string Description {  get; set; } = string.Empty;
-
-        public List<Product> Products { get; set; }
-    }
-}
