--- Data/Enums/Gender.cs
@@ -1,8 +1,0 @@
-﻿namespace Data.Enums
-{
-    public enum Gender
-    {
-        Male =1,Female =2
-    }
-
-}
