--- Data/Enums/Role.cs
@@ -1,9 +1,0 @@
-﻿namespace Data.Enums
-{
-    public  enum Role
-    {
-        User= 1,
-        Admin =2,
-        SuperAdmin =3
-    }
-}
