--- Data/Enums/Status.cs
@@ -1,12 +1,0 @@
-﻿
-
-namespace Data.Enums
-{
-    public  enum Status
-    {
-        todo = 1,
-        inprocess  = 2,
-        done = 3,
-
-    }
-}
