+++ Domain/Entities/Category.cs
@@ -1,0 +1,11 @@
+﻿using Data.Entities;
+
+namespace Domain.Entities;
+
+public class Category : BaseEntity
+{
+    public string Name { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
+
+    public List<Product> Products { get; set; }
+}
