+++ Domain/Enums/Gender.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Enums;
+
+public enum Gender
+{
+    Male = 1,
+    Female = 2
+}
