+++ Domain/Enums/Role.cs
@@ -1,0 +1,8 @@
+﻿namespace Domain.Enums;
+
+public enum Role
+{
+    User = 1,
+    Admin = 2,
+    SuperAdmin = 3
+}
