+++ Domain/Enums/Satus.cs
@@ -1,0 +1,8 @@
+﻿namespace Domain.Enums;
+
+public enum Status
+{
+    todo = 1,
+    inprocess = 2,
+    done = 3,
+}
